refactor (imports): use local path internally (not exports from index.ts)
