+++ src/api/admin/permissions/permissions-endpoint.spec.ts
@@ -1,4 +1,4 @@
-import { ApiResponseData, ApiResponseError, UsersResponse } from "../../..";
+import { ApiResponseData } from "../../../models/api-response-data";
 import { MockAjaxCall } from "../../../../test/mockajaxcall";
 import { KnoraApiConfig } from "../../../knora-api-config";
 import { KnoraApiConnection } from "../../../knora-api-connection";
