+++ src/api/v2/resource/resources-endpoint.ts
@@ -1,7 +1,8 @@
 import { Observable } from "rxjs";
 import { AjaxResponse } from "rxjs/ajax";
 import { catchError, map, mergeMap } from "rxjs/operators";
-import { ApiResponseError, KnoraApiConfig } from "../../..";
+import { ApiResponseError } from "../../../models/api-response-error";
+import { KnoraApiConfig } from "../../../knora-api-config";
 import { ReadResource } from "../../../models/v2/resources/read-resource";
 import { ResourcesConversionUtil } from "../../../models/v2/resources/ResourcesConversionUtil";
 import { Endpoint } from "../../endpoint";
